+++ concourse-integration-tests/src/test/java/org/cinchapi/concourse/ConcourseIntegrationTest.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  * 
- * Copyright (c) 2013-2014 Jeff Nelson, Cinchapi Software Collective
+ * Copyright (c) 2013-2015 Jeff Nelson, Cinchapi Software Collective
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -100,6 +100,8 @@ public abstract class ConcourseIntegrationTest {
             System.out.println("---");
             System.out.println(Variables.dump());
             System.out.println("");
+            stop();
+            afterEachTest();
         }
 
         @Override