+++ concourse-server/src/main/java/org/cinchapi/concourse/server/storage/Stores.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  * 
- * Copyright (c) 2014 Jeff Nelson, Cinchapi Software Collective
+ * Copyright (c) 2013-2015 Jeff Nelson, Cinchapi Software Collective
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -27,6 +27,7 @@ import org.cinchapi.concourse.Link;
 import org.cinchapi.concourse.thrift.Operator;
 import org.cinchapi.concourse.thrift.TObject;
 import org.cinchapi.concourse.util.Convert;
+import org.cinchapi.concourse.util.TStrings;
 
 /**
  * {@link Store} based utility functions.
@@ -44,7 +45,16 @@ public final class Stores {
      * @return the normalized Operator
      */
     public static Operator normalizeOperator(Operator operator) {
-        return operator == Operator.LINKS_TO ? Operator.EQUALS : operator;
+    	switch (operator) {
+		case LIKE:
+			return Operator.REGEX;
+		case NOT_LIKE:
+			return Operator.NOT_REGEX;
+		case LINKS_TO:
+			return Operator.EQUALS;
+		default:
+			return operator;
+		}
     }
 
     /**
@@ -57,13 +67,28 @@ public final class Stores {
      */
     public static TObject normalizeValue(Operator operator, TObject value) {
         try {
-            return operator == Operator.LINKS_TO ? Convert.javaToThrift(Link
-                    .to(((Number) Convert.thriftToJava(value)).longValue()))
-                    : value;
+            switch (operator) {
+            case REGEX:
+            case NOT_REGEX:
+                value = Convert.javaToThrift(((String) Convert
+                        .thriftToJava(value)).replaceAll(
+                        TStrings.REGEX_PERCENT_SIGN_WITHOUT_ESCAPE_CHAR, ".*")
+                        .replaceAll(
+                                TStrings.REGEX_PERCENT_SIGN_WITH_ESCAPE_CHAR,
+                                "%"));
+                break;
+            case LINKS_TO:
+                value = Convert.javaToThrift(Link.to(((Number) Convert
+                        .thriftToJava(value)).longValue()));
+                break;
+            default:
+                // noop: default case added to suppress compiler warning
+                break;
+            }
+            return value;
         }
         catch (ClassCastException e) {
             return value;
         }
     }
-
 }