+++ concourse-server/src/main/java/org/cinchapi/concourse/server/storage/cache/BloomFilter.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  * 
- * Copyright (c) 2013-2014 Jeff Nelson, Cinchapi Software Collective
+ * Copyright (c) 2013-2015 Jeff Nelson, Cinchapi Software Collective
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -96,7 +96,7 @@ public class BloomFilter implements Syncable {
      * @param file
      * @return the BloomFilter
      */
-    @SuppressWarnings({ "unchecked", "resource" })
+    @SuppressWarnings({ "unchecked" })
     public static BloomFilter open(String file) {
         try {
             final AtomicBoolean upgrade = new AtomicBoolean(false);